requirements.txt
web3
gmpy2
loguru
Pillow
requests
pycrypto
pwntools
py-solc-x